Changed the prompt for headpieces
The previous prompt resulted in most images just being lame. I have
experimented for a while and the most general approach that seems to
work at least sometimes is to restrict the overal theme. Say, 'historic
photos' work well. In the end, going with 'human activity'. That
obviously won't suit every book, but at least the results are a bit more
to my liking.